+++ src/main/java/net/journey/entity/mob/boiling/EntityMagmaBlaze.java
@@ -74,7 +74,7 @@ public class EntityMagmaBlaze extends JEntityMob {
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataManager.register(ON_FIRE, Byte.valueOf((byte) 0));
+        this.dataManager.register(ON_FIRE, (byte) 0);
     }
 
     @SideOnly(Side.CLIENT)
@@ -161,11 +161,11 @@ public class EntityMagmaBlaze extends JEntityMob {
     }
 
     public boolean isFlying() {
-        return (this.dataManager.get(ON_FIRE).byteValue() & 1) != 0;
+        return (this.dataManager.get(ON_FIRE) & 1) != 0;
     }
 
     public void setOnFire(boolean onFire) {
-        byte b0 = this.dataManager.get(ON_FIRE).byteValue();
+        byte b0 = this.dataManager.get(ON_FIRE);
 
         if (onFire) {
             b0 = (byte) (b0 | 1);
@@ -173,7 +173,7 @@ public class EntityMagmaBlaze extends JEntityMob {
             b0 = (byte) (b0 & -2);
         }
 
-        this.dataManager.set(ON_FIRE, Byte.valueOf(b0));
+        this.dataManager.set(ON_FIRE, b0);
     }
 
     @Override
