+++ src/main/java/net/journey/entity/mob/boiling/EntityScreamer.java
@@ -70,7 +70,7 @@ public class EntityScreamer extends JEntityMob {
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataManager.register(ON_FIRE, Byte.valueOf((byte) 0));
+        this.dataManager.register(ON_FIRE, (byte) 0);
     }
 
     @Override
@@ -147,14 +147,14 @@ public class EntityScreamer extends JEntityMob {
     }
 
     public boolean isFlying() {
-        return (this.dataManager.get(ON_FIRE).byteValue() & 1) != 0;
+        return (this.dataManager.get(ON_FIRE) & 1) != 0;
     }
 
     public void setFlying(boolean b) {
-        byte b0 = this.dataManager.get(ON_FIRE).byteValue();
+        byte b0 = this.dataManager.get(ON_FIRE);
         if (b) b0 = (byte) (b0 | 1);
         else b0 &= -2;
-        this.dataManager.set(ON_FIRE, Byte.valueOf(b0));
+        this.dataManager.set(ON_FIRE, b0);
     }
 
     @Override
