+++ src/main/java/net/journey/entity/mob/terrania/mob/EntityPurplian.java
@@ -66,7 +66,7 @@ public class EntityPurplian extends JEntityMob {
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataManager.register(ON_FIRE, Byte.valueOf((byte) 0));
+        this.dataManager.register(ON_FIRE, (byte) 0);
     }
 
     @Override
@@ -137,14 +137,14 @@ public class EntityPurplian extends JEntityMob {
     }
 
     public boolean isFlying() {
-        return (this.dataManager.get(ON_FIRE).byteValue() & 1) != 0;
+        return (this.dataManager.get(ON_FIRE) & 1) != 0;
     }
 
     public void setFlying(boolean b) {
-        byte b0 = this.dataManager.get(ON_FIRE).byteValue();
+        byte b0 = this.dataManager.get(ON_FIRE);
         if (b) b0 = (byte) (b0 | 1);
         else b0 &= -2;
-        this.dataManager.set(ON_FIRE, Byte.valueOf(b0));
+        this.dataManager.set(ON_FIRE, b0);
     }
 
     class AIFireballAttack extends EntityAIBase {
